requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
scikit-learn>=1.0.2
lightgbm>=3.3.0
joblib>=1.0.1
PyYAML>=5.4.1
numpy>=1.21.0